clickreviews/cr_desktop: fix check on icon extension
The check was failing even if a file with the supported extension
existed. To fix it, use a combination of map() and any(): the former
maps the list of the icon extensions to a boolean telling if an icon
with that extension was found; the latter is used to check if any of
these checks returned true.